Refactor to move RequestSourcePlugins to a top level RequestSource folder. Signed-off-by: William Juan <66322422+wjuan-AFK@users.noreply.github.com>
