+++ source/common/BUILD
@@ -82,30 +82,6 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "request_source_plugin_impl",
-    srcs = [
-        "request_source_plugin_impl.cc",
-    ],
-    hdrs = [
-        "request_source_plugin_impl.h",
-    ],
-    repository = "@envoy",
-    visibility = ["//visibility:public"],
-    deps = [
-        ":nighthawk_common_lib",
-        ":request_impl_lib",
-        ":request_source_impl_lib",
-        "//include/nighthawk/common:request_source_plugin_config_factory_lib",
-        "@envoy//source/common/common:thread_lib_with_external_headers",
-        "@envoy//source/common/protobuf:message_validator_lib_with_external_headers",
-        "@envoy//source/common/protobuf:protobuf_with_external_headers",
-        "@envoy//source/common/protobuf:utility_lib_with_external_headers",
-        "@envoy//source/exe:platform_header_lib_with_external_headers",
-        "@envoy//source/exe:platform_impl_lib",
-    ],
-)
-
-envoy_cc_library(
     name = "version_linkstamp",
     srcs = ["version_linkstamp.cc"],
     repository = "@envoy",
